hierarchy_RL/requirements.txt
torch>=1.9.0
numpy>=1.21.0
gymnasium>=0.28.0
tqdm>=4.62.0
matplotlib>=3.5.0
Pillow>=8.3.0
pygame>=2.1.0
